# Conflicts:
#	QMChatService/QMChatService.xcodeproj/project.pbxproj
#	QMChatService/QMChatService/QMAttachmentService/QMChatAttachmentService.m
#	QMChatService/QMChatService/QMChatService.h
#	QMChatService/QMChatService/QMChatService.m
#	QMChatService/QMChatService/QMMediaService/Services/QMMediaInfoService/QMAsynchronousOperation.m
#	QMChatService/QMChatService/QMMediaService/Services/QMMediaInfoService/QMImageOperation.m
#	QMServices/QMServices.h

+++ QMChatCache/QMChatCache/CoreData/Entries/CDMessage.m
@@ -47,7 +47,7 @@
     self.delayedValue = message.delayed;
     self.text = message.text;
     self.dateSend = message.dateSent;
-    self.recipientIDValue = message.recipientID;
+    self.recipientIDValue = (int32_t)message.recipientID;
     self.senderID = @(message.senderID);
     self.dialogID = message.dialogID;
     self.customParameters = [self binaryDataWithObject:message.customParameters];
